Pause Menu Fix
When the pause menu is initially replaced it was not being made a child of the player, as this was based on an old version.
Now it becomes a child of the player. This prevents accidentally having two menus in existence at once.
This fixed a bug where you had to press escape twice to open the menu once MP is started.