Only call player ready when its the local player
Before when you try to save as host you would get the error: Attempt to call function 'save' in base 'previously freed' on a null instance.

because the newly created client player instance made itself the new player node in the cogito_scene_manager of the host in the normal _ready() function. Then when saving as the host it would try to save the current_player_interaction_component of this non-local player node, which would have been freed in the _enter_tree() function because it is not the local player node.

So the solution I propose is to only call the normal _ready() function if it is the local player.